fix: PR 피드백 반영하여 axios 요청 분리
